Working with multiple tables in mvc
working with 2 related tables in MVC
1. tblDepartment
2. tblEmployee

This is what we want to achieve:
1. Display all the departments from tblDepartments table. The Department
names should be rendered as hyperlinks.
2. On clicking the department name link, all the employees in the
department should be displayed. The employee names should be rendered as
hyperlinks.
3. On clicking the employee name link, the full details of the employee
should be displayed.
4. A link should also be provided on the employee full details page to
navigate back to Employee list page. Along the same lines, a link should
also be provided on the employee list page to navigate back to
Departments list page.